requirements.txt
attrs==21.2.0
cachelib==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-Session==0.4.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
redis==3.5.3
requests==2.26.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
